+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/application/ApplicationScreen.kt
@@ -20,86 +20,52 @@ package com.eblan.launcher.feature.home.screen.application
 import android.graphics.Rect
 import android.os.Build
 import android.os.UserHandle
-import androidx.activity.compose.BackHandler
 import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.compose.animation.SharedTransitionScope
 import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.gestures.detectVerticalDragGestures
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.calculateEndPadding
 import androidx.compose.foundation.layout.calculateStartPadding
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.isImeVisible
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
-import androidx.compose.foundation.lazy.grid.rememberLazyGridState
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.pager.HorizontalPager
-import androidx.compose.foundation.pager.rememberPagerState
-import androidx.compose.foundation.rememberOverscrollEffect
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.input.TextFieldState
-import androidx.compose.foundation.text.input.clearText
-import androidx.compose.foundation.text.input.rememberTextFieldState
 import androidx.compose.material3.ElevatedFilterChip
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FilterChipDefaults
-import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.SearchBar
-import androidx.compose.material3.SearchBarDefaults
-import androidx.compose.material3.SearchBarState
-import androidx.compose.material3.SearchBarValue
 import androidx.compose.material3.SecondaryTabRow
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Tab
 import androidx.compose.material3.Text
-import androidx.compose.material3.rememberSearchBarState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.key
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.drawWithContent
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.layer.drawLayer
 import androidx.compose.ui.graphics.rememberGraphicsLayer
-import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.layout.positionInRoot
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.Dp
@@ -114,43 +80,33 @@ import coil3.request.ImageRequest
 import coil3.request.addLastModifiedToFileCacheKey
 import com.eblan.launcher.designsystem.icon.EblanLauncherIcons
 import com.eblan.launcher.domain.model.AppDrawerSettings
+import com.eblan.launcher.domain.model.AppDrawerType
 import com.eblan.launcher.domain.model.Associate
 import com.eblan.launcher.domain.model.EblanAction
 import com.eblan.launcher.domain.model.EblanActionType
 import com.eblan.launcher.domain.model.EblanAppWidgetProviderInfo
 import com.eblan.launcher.domain.model.EblanApplicationInfo
 import com.eblan.launcher.domain.model.EblanApplicationInfoGroup
-import com.eblan.launcher.domain.model.EblanApplicationInfoOrder
 import com.eblan.launcher.domain.model.EblanApplicationInfoTag
 import com.eblan.launcher.domain.model.EblanShortcutInfo
 import com.eblan.launcher.domain.model.EblanShortcutInfoByGroup
-import com.eblan.launcher.domain.model.EblanUser
 import com.eblan.launcher.domain.model.EblanUserPageKey
-import com.eblan.launcher.domain.model.EblanUserType
 import com.eblan.launcher.domain.model.GetEblanApplicationInfosByLabel
 import com.eblan.launcher.domain.model.GridItem
 import com.eblan.launcher.domain.model.GridItemData
 import com.eblan.launcher.domain.model.ManagedProfileResult
 import com.eblan.launcher.domain.model.TextColor
-import com.eblan.launcher.feature.home.component.scroll.OffsetNestedScrollConnection
-import com.eblan.launcher.feature.home.component.scroll.OffsetOverscrollEffect
-import com.eblan.launcher.feature.home.dialog.EblanApplicationInfoOrderDialog
 import com.eblan.launcher.feature.home.model.Drag
 import com.eblan.launcher.feature.home.model.GridItemSource
 import com.eblan.launcher.feature.home.model.SharedElementKey
-import com.eblan.launcher.feature.home.screen.application.draganddrop.DragAndDropEblanApplicationInfos
+import com.eblan.launcher.feature.home.screen.application.horizontal.HorizontalApplicationScreen
+import com.eblan.launcher.feature.home.screen.application.vertical.VerticalApplicationScreen
 import com.eblan.launcher.feature.home.util.getHorizontalAlignment
 import com.eblan.launcher.feature.home.util.getSystemTextColor
 import com.eblan.launcher.feature.home.util.getVerticalArrangement
 import com.eblan.launcher.framework.packagemanager.AndroidPackageManagerWrapper
 import com.eblan.launcher.framework.usermanager.AndroidUserManagerWrapper
 import com.eblan.launcher.ui.local.LocalLauncherApps
-import com.eblan.launcher.ui.local.LocalPackageManager
-import com.eblan.launcher.ui.local.LocalUserManager
-import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.debounce
-import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 import kotlin.math.roundToInt
 import kotlin.uuid.ExperimentalUuidApi
@@ -227,590 +183,84 @@ internal fun SharedTransitionScope.ApplicationScreen(
             }
         },
     ) {
-        Success(
-            appDrawerSettings = appDrawerSettings,
-            currentPage = currentPage,
-            drag = drag,
-            eblanAppWidgetProviderInfosGroup = eblanAppWidgetProviderInfosGroup,
-            eblanApplicationInfoTags = eblanApplicationInfoTags,
-            eblanShortcutInfosGroup = eblanShortcutInfosGroup,
-            getEblanApplicationInfosByLabel = getEblanApplicationInfosByLabel,
-            hasShortcutHostPermission = hasShortcutHostPermission,
-            iconPackFilePaths = iconPackFilePaths,
-            isPressHome = isPressHome,
-            managedProfileResult = managedProfileResult,
-            paddingValues = paddingValues,
-            screenHeight = screenHeight,
-            swipeY = swipeY,
-            isVisibleOverlay = isVisibleOverlay,
-            onDismiss = onDismiss,
-            onDragEnd = onDragEnd,
-            onDraggingGridItem = onDraggingGridItem,
-            onEditApplicationInfo = onEditApplicationInfo,
-            onGetEblanApplicationInfosByLabel = onGetEblanApplicationInfosByLabel,
-            onGetEblanApplicationInfosByTagId = onGetEblanApplicationInfosByTagId,
-            onUpdateAppDrawerSettings = onUpdateAppDrawerSettings,
-            onUpdateEblanApplicationInfos = onUpdateEblanApplicationInfos,
-            onUpdateGridItemSource = onUpdateGridItemSource,
-            onUpdateImageBitmap = onUpdateImageBitmap,
-            onUpdateIsDragging = onUpdateIsDragging,
-            onUpdateOverlayBounds = onUpdateOverlayBounds,
-            onUpdateSharedElementKey = onUpdateSharedElementKey,
-            onVerticalDrag = onVerticalDrag,
-            onWidgets = onWidgets,
-            onDraggingShortcutInfoGridItem = onDraggingShortcutInfoGridItem,
-            onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class, ExperimentalSharedTransitionApi::class, FlowPreview::class)
-@Composable
-private fun SharedTransitionScope.Success(
-    modifier: Modifier = Modifier,
-    appDrawerSettings: AppDrawerSettings,
-    currentPage: Int,
-    drag: Drag,
-    eblanAppWidgetProviderInfosGroup: Map<String, List<EblanAppWidgetProviderInfo>>,
-    eblanApplicationInfoTags: List<EblanApplicationInfoTag>,
-    eblanShortcutInfosGroup: Map<EblanShortcutInfoByGroup, List<EblanShortcutInfo>>,
-    getEblanApplicationInfosByLabel: GetEblanApplicationInfosByLabel,
-    hasShortcutHostPermission: Boolean,
-    iconPackFilePaths: Map<String, String>,
-    isPressHome: Boolean,
-    managedProfileResult: ManagedProfileResult?,
-    paddingValues: PaddingValues,
-    screenHeight: Int,
-    swipeY: Float,
-    isVisibleOverlay: Boolean,
-    onDismiss: () -> Unit,
-    onDragEnd: (Float) -> Unit,
-    onDraggingGridItem: () -> Unit,
-    onEditApplicationInfo: (
-        serialNumber: Long,
-        componentName: String,
-    ) -> Unit,
-    onGetEblanApplicationInfosByLabel: (String) -> Unit,
-    onGetEblanApplicationInfosByTagId: (Long?) -> Unit,
-    onUpdateAppDrawerSettings: (AppDrawerSettings) -> Unit,
-    onUpdateEblanApplicationInfos: (List<EblanApplicationInfo>) -> Unit,
-    onUpdateGridItemSource: (GridItemSource) -> Unit,
-    onUpdateImageBitmap: (ImageBitmap) -> Unit,
-    onUpdateIsDragging: (Boolean) -> Unit,
-    onUpdateOverlayBounds: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
-    onVerticalDrag: (Float) -> Unit,
-    onWidgets: (EblanApplicationInfoGroup) -> Unit,
-    onDraggingShortcutInfoGridItem: () -> Unit,
-    onUpdateIsVisibleOverlay: (Boolean) -> Unit,
-) {
-    val density = LocalDensity.current
-
-    var showPopupApplicationMenu by remember { mutableStateOf(false) }
-
-    var popupIntOffset by remember { mutableStateOf(IntOffset.Zero) }
-
-    var popupIntSize by remember { mutableStateOf(IntSize.Zero) }
-
-    val launcherApps = LocalLauncherApps.current
-
-    val leftPadding = with(density) {
-        paddingValues.calculateStartPadding(LayoutDirection.Ltr).roundToPx()
-    }
-
-    val topPadding = with(density) {
-        paddingValues.calculateTopPadding().roundToPx()
-    }
-
-    val horizontalPagerState = rememberPagerState(
-        pageCount = {
-            getEblanApplicationInfosByLabel.eblanApplicationInfos.keys.size
-        },
-    )
-
-    val appDrawerRowsHeight = with(density) {
-        appDrawerSettings.appDrawerRowsHeight.dp.roundToPx()
-    }
-
-    val searchBarState = rememberSearchBarState()
-
-    val textFieldState = rememberTextFieldState()
-
-    var selectedEblanApplicationInfoTagId by remember { mutableStateOf<Long?>(null) }
-
-    var isRearrangeEblanApplicationInfo by remember { mutableStateOf(false) }
-
-    var showEblanApplicationInfoOrderDialog by remember { mutableStateOf(false) }
-
-    var selectedEblanApplicationInfo by remember { mutableStateOf<EblanApplicationInfo?>(null) }
-
-    LaunchedEffect(key1 = textFieldState) {
-        snapshotFlow { textFieldState.text }.debounce(500L).onEach { text ->
-            onGetEblanApplicationInfosByLabel(text.toString())
-
-            showPopupApplicationMenu = false
-        }.collect()
-    }
-
-    LaunchedEffect(key1 = swipeY) {
-        if (swipeY.roundToInt() >= screenHeight && textFieldState.text.isNotEmpty()) {
-            onGetEblanApplicationInfosByLabel("")
-
-            textFieldState.clearText()
-
-            selectedEblanApplicationInfoTagId = null
-        }
-
-        if (swipeY.roundToInt() > 0 && showPopupApplicationMenu) {
-            showPopupApplicationMenu = false
-        }
-    }
-
-    LaunchedEffect(key1 = Unit) {
-        snapshotFlow { selectedEblanApplicationInfoTagId }.onEach { selectedEblanApplicationInfoTag ->
-            onGetEblanApplicationInfosByTagId(selectedEblanApplicationInfoTag)
-        }.collect()
-    }
-
-    LaunchedEffect(key1 = isPressHome) {
-        if (isPressHome) {
-            showPopupApplicationMenu = false
-
-            searchBarState.animateToCollapsed()
-
-            onDismiss()
-        }
-    }
-
-    LaunchedEffect(key1 = drag) {
-        if (drag == Drag.Start && searchBarState.currentValue == SearchBarValue.Expanded) {
-            searchBarState.animateToCollapsed()
-        }
-    }
-
-    LaunchedEffect(key1 = horizontalPagerState.isScrollInProgress) {
-        if (horizontalPagerState.isScrollInProgress && showPopupApplicationMenu) {
-            showPopupApplicationMenu = false
-        }
-    }
-
-    BackHandler(enabled = swipeY < screenHeight.toFloat()) {
-        showPopupApplicationMenu = false
-
-        onDismiss()
-    }
-
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .padding(
-                top = paddingValues.calculateTopPadding(),
-                start = paddingValues.calculateStartPadding(LayoutDirection.Ltr),
-                end = paddingValues.calculateEndPadding(LayoutDirection.Ltr),
-            ),
-    ) {
-        ApplicationSearchBar(
-            searchBarState = searchBarState,
-            textFieldState = textFieldState,
-            swipeY = swipeY,
-            showKeyboard = appDrawerSettings.showKeyboard,
-            onUpdateShowEblanApplicationInfoOrderDialog = { newShowEblanApplicationInfoOrderDialog ->
-                showEblanApplicationInfoOrderDialog = newShowEblanApplicationInfoOrderDialog
-            },
-        )
-
-        if (eblanApplicationInfoTags.isNotEmpty()) {
-            LazyRow(modifier = Modifier.fillMaxWidth()) {
-                items(eblanApplicationInfoTags) { eblanApplicationInfoTag ->
-                    TagElevatedFilterChip(
-                        eblanApplicationInfoTag = eblanApplicationInfoTag,
-                        selectedEblanApplicationInfoTag = selectedEblanApplicationInfoTagId,
-                        onUpdateEblanApplicationInfoTag = { newEblanApplicationInfoTagId ->
-                            selectedEblanApplicationInfoTagId = newEblanApplicationInfoTagId
-                        },
-                    )
-                }
-            }
-        }
-
-        if (getEblanApplicationInfosByLabel.eblanApplicationInfos.keys.size > 1) {
-            EblanApplicationInfoTabRow(
-                currentPage = horizontalPagerState.currentPage,
-                eblanApplicationInfos = getEblanApplicationInfosByLabel.eblanApplicationInfos,
-                onAnimateScrollToPage = horizontalPagerState::animateScrollToPage,
-            )
-
-            HorizontalPager(
-                modifier = Modifier.fillMaxSize(),
-                state = horizontalPagerState,
-            ) { index ->
-                EblanApplicationInfosPage(
+        when (appDrawerSettings.appDrawerType) {
+            AppDrawerType.Vertical -> {
+                VerticalApplicationScreen(
                     appDrawerSettings = appDrawerSettings,
                     currentPage = currentPage,
                     drag = drag,
-                    eblanApplicationInfoOrder = appDrawerSettings.eblanApplicationInfoOrder,
+                    eblanAppWidgetProviderInfosGroup = eblanAppWidgetProviderInfosGroup,
+                    eblanApplicationInfoTags = eblanApplicationInfoTags,
+                    eblanShortcutInfosGroup = eblanShortcutInfosGroup,
                     getEblanApplicationInfosByLabel = getEblanApplicationInfosByLabel,
+                    hasShortcutHostPermission = hasShortcutHostPermission,
                     iconPackFilePaths = iconPackFilePaths,
-                    index = index,
-                    isRearrangeEblanApplicationInfo = isRearrangeEblanApplicationInfo,
+                    isPressHome = isPressHome,
                     managedProfileResult = managedProfileResult,
                     paddingValues = paddingValues,
+                    screenHeight = screenHeight,
+                    swipeY = swipeY,
                     isVisibleOverlay = isVisibleOverlay,
-                    showPopupApplicationMenu = showPopupApplicationMenu,
                     onDismiss = onDismiss,
-                    onDismissDragAndDrop = {
-                        isRearrangeEblanApplicationInfo = false
-                    },
                     onDragEnd = onDragEnd,
                     onDraggingGridItem = onDraggingGridItem,
+                    onEditApplicationInfo = onEditApplicationInfo,
+                    onGetEblanApplicationInfosByLabel = onGetEblanApplicationInfosByLabel,
+                    onGetEblanApplicationInfosByTagId = onGetEblanApplicationInfosByTagId,
+                    onUpdateAppDrawerSettings = onUpdateAppDrawerSettings,
                     onUpdateEblanApplicationInfos = onUpdateEblanApplicationInfos,
                     onUpdateGridItemSource = onUpdateGridItemSource,
                     onUpdateImageBitmap = onUpdateImageBitmap,
                     onUpdateIsDragging = onUpdateIsDragging,
-                    onUpdateOverlayBounds = { intOffset, intSize ->
-                        onUpdateOverlayBounds(intOffset, intSize)
-
-                        popupIntOffset = intOffset
-
-                        popupIntSize = intSize
-                    },
-                    onUpdatePopupMenu = { newShowPopupApplicationMenu ->
-                        showPopupApplicationMenu = newShowPopupApplicationMenu
-                    },
+                    onUpdateOverlayBounds = onUpdateOverlayBounds,
                     onUpdateSharedElementKey = onUpdateSharedElementKey,
                     onVerticalDrag = onVerticalDrag,
-                    onUpdateEblanApplicationInfo = { eblanApplicationInfo ->
-                        selectedEblanApplicationInfo = eblanApplicationInfo
-                    },
+                    onWidgets = onWidgets,
+                    onDraggingShortcutInfoGridItem = onDraggingShortcutInfoGridItem,
                     onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
                 )
             }
-        } else {
-            EblanApplicationInfosPage(
-                appDrawerSettings = appDrawerSettings,
-                currentPage = currentPage,
-                drag = drag,
-                eblanApplicationInfoOrder = appDrawerSettings.eblanApplicationInfoOrder,
-                getEblanApplicationInfosByLabel = getEblanApplicationInfosByLabel,
-                iconPackFilePaths = iconPackFilePaths,
-                index = 0,
-                isRearrangeEblanApplicationInfo = isRearrangeEblanApplicationInfo,
-                managedProfileResult = managedProfileResult,
-                paddingValues = paddingValues,
-                isVisibleOverlay = isVisibleOverlay,
-                showPopupApplicationMenu = showPopupApplicationMenu,
-                onDismiss = onDismiss,
-                onDismissDragAndDrop = {
-                    isRearrangeEblanApplicationInfo = false
-                },
-                onDragEnd = onDragEnd,
-                onDraggingGridItem = onDraggingGridItem,
-                onUpdateEblanApplicationInfos = onUpdateEblanApplicationInfos,
-                onUpdateGridItemSource = onUpdateGridItemSource,
-                onUpdateImageBitmap = onUpdateImageBitmap,
-                onUpdateIsDragging = onUpdateIsDragging,
-                onUpdateOverlayBounds = { intOffset, intSize ->
-                    onUpdateOverlayBounds(intOffset, intSize)
 
-                    popupIntOffset = intOffset
-
-                    popupIntSize = IntSize(
-                        width = intSize.width,
-                        height = appDrawerRowsHeight,
-                    )
-                },
-                onUpdatePopupMenu = { newShowPopupApplicationMenu ->
-                    showPopupApplicationMenu = newShowPopupApplicationMenu
-                },
-                onUpdateSharedElementKey = onUpdateSharedElementKey,
-                onVerticalDrag = onVerticalDrag,
-                onUpdateEblanApplicationInfo = { eblanApplicationInfo ->
-                    selectedEblanApplicationInfo = eblanApplicationInfo
-                },
-                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-            )
-        }
-    }
-
-    if (showPopupApplicationMenu && selectedEblanApplicationInfo != null) {
-        ApplicationInfoPopup(
-            currentPage = currentPage,
-            drag = drag,
-            eblanAppWidgetProviderInfos = eblanAppWidgetProviderInfosGroup,
-            eblanShortcutInfosGroup = eblanShortcutInfosGroup,
-            eblanApplicationInfo = selectedEblanApplicationInfo,
-            gridItemSettings = appDrawerSettings.gridItemSettings,
-            hasShortcutHostPermission = hasShortcutHostPermission,
-            paddingValues = paddingValues,
-            popupIntOffset = popupIntOffset,
-            popupIntSize = popupIntSize,
-            onDismissRequest = {
-                showPopupApplicationMenu = false
-            },
-            onDraggingShortcutInfoGridItem = {
-                showPopupApplicationMenu = false
-
-                onDismiss()
-
-                onDraggingShortcutInfoGridItem()
-            },
-            onEditApplicationInfo = onEditApplicationInfo,
-            onTapShortcutInfo = { serialNumber, packageName, shortcutId ->
-                val sourceBoundsX = popupIntOffset.x + leftPadding
-
-                val sourceBoundsY = popupIntOffset.y + topPadding
-
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N_MR1) {
-                    launcherApps.startShortcut(
-                        serialNumber = serialNumber,
-                        packageName = packageName,
-                        id = shortcutId,
-                        sourceBounds = Rect(
-                            sourceBoundsX,
-                            sourceBoundsY,
-                            sourceBoundsX + popupIntSize.width,
-                            sourceBoundsY + popupIntSize.height,
-                        ),
-                    )
-                }
-            },
-            onUpdateGridItemSource = onUpdateGridItemSource,
-            onUpdateImageBitmap = onUpdateImageBitmap,
-            onUpdateOverlayBounds = onUpdateOverlayBounds,
-            onUpdateSharedElementKey = onUpdateSharedElementKey,
-            onWidgets = onWidgets,
-            onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-        )
-    }
-
-    if (showEblanApplicationInfoOrderDialog) {
-        EblanApplicationInfoOrderDialog(
-            eblanApplicationInfoOrder = appDrawerSettings.eblanApplicationInfoOrder,
-            onDismissRequest = {
-                showEblanApplicationInfoOrderDialog = false
-            },
-            onUpdateClick = { eblanApplicationInfoOrder, newIsRearrangeEblanApplicationInfo ->
-                onUpdateAppDrawerSettings(appDrawerSettings.copy(eblanApplicationInfoOrder = eblanApplicationInfoOrder))
-
-                isRearrangeEblanApplicationInfo = newIsRearrangeEblanApplicationInfo
-
-                showEblanApplicationInfoOrderDialog = false
-            },
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun ApplicationSearchBar(
-    modifier: Modifier = Modifier,
-    searchBarState: SearchBarState,
-    textFieldState: TextFieldState,
-    swipeY: Float,
-    showKeyboard: Boolean,
-    onUpdateShowEblanApplicationInfoOrderDialog: (Boolean) -> Unit,
-) {
-    val keyboardController = LocalSoftwareKeyboardController.current
-
-    val scope = rememberCoroutineScope()
-
-    val focusRequester = remember { FocusRequester() }
-
-    LaunchedEffect(key1 = swipeY) {
-        if (swipeY.roundToInt() == 0 && showKeyboard) {
-            focusRequester.requestFocus()
-
-            keyboardController?.show()
-        }
-    }
-
-    SearchBar(
-        state = searchBarState,
-        modifier = modifier
-            .focusRequester(focusRequester)
-            .fillMaxWidth()
-            .padding(10.dp),
-        inputField = {
-            SearchBarDefaults.InputField(
-                textFieldState = textFieldState,
-                searchBarState = searchBarState,
-                leadingIcon = {
-                    Icon(
-                        imageVector = EblanLauncherIcons.Search,
-                        contentDescription = null,
-                    )
-                },
-                trailingIcon = {
-                    IconButton(
-                        onClick = {
-                            onUpdateShowEblanApplicationInfoOrderDialog(true)
-                        },
-                    ) {
-                        Icon(
-                            imageVector = EblanLauncherIcons.MoreVert,
-                            contentDescription = null,
-                        )
-                    }
-                },
-                onSearch = { scope.launch { searchBarState.animateToCollapsed() } },
-                placeholder = { Text(text = "Search Applications") },
-            )
-        },
-    )
-}
-
-@OptIn(ExperimentalLayoutApi::class, ExperimentalSharedTransitionApi::class)
-@Composable
-private fun SharedTransitionScope.EblanApplicationInfosPage(
-    modifier: Modifier = Modifier,
-    appDrawerSettings: AppDrawerSettings,
-    currentPage: Int,
-    drag: Drag,
-    eblanApplicationInfoOrder: EblanApplicationInfoOrder,
-    getEblanApplicationInfosByLabel: GetEblanApplicationInfosByLabel,
-    iconPackFilePaths: Map<String, String>,
-    index: Int,
-    isRearrangeEblanApplicationInfo: Boolean,
-    managedProfileResult: ManagedProfileResult?,
-    paddingValues: PaddingValues,
-    showPopupApplicationMenu: Boolean,
-    isVisibleOverlay: Boolean,
-    onDismiss: () -> Unit,
-    onDismissDragAndDrop: () -> Unit,
-    onDragEnd: (Float) -> Unit,
-    onDraggingGridItem: () -> Unit,
-    onUpdateEblanApplicationInfos: (List<EblanApplicationInfo>) -> Unit,
-    onUpdateGridItemSource: (GridItemSource) -> Unit,
-    onUpdateImageBitmap: (ImageBitmap) -> Unit,
-    onUpdateIsDragging: (Boolean) -> Unit,
-    onUpdateOverlayBounds: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onUpdatePopupMenu: (Boolean) -> Unit,
-    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
-    onVerticalDrag: (Float) -> Unit,
-    onUpdateEblanApplicationInfo: (EblanApplicationInfo) -> Unit,
-    onUpdateIsVisibleOverlay: (Boolean) -> Unit,
-) {
-    val userManager = LocalUserManager.current
-
-    val packageManager = LocalPackageManager.current
-
-    val eblanUserPageKey =
-        getEblanApplicationInfosByLabel.eblanApplicationInfos.keys.toList().getOrElse(
-            index = index,
-            defaultValue = {
-                EblanUserPageKey(
-                    eblanUser = EblanUser(
-                        serialNumber = 0L,
-                        eblanUserType = EblanUserType.Personal,
-                        isPrivateSpaceEntryPointHidden = false,
-                    ),
-                    page = 0,
+            AppDrawerType.Horizontal -> {
+                HorizontalApplicationScreen(
+                    appDrawerSettings = appDrawerSettings,
+                    currentPage = currentPage,
+                    drag = drag,
+                    eblanAppWidgetProviderInfosGroup = eblanAppWidgetProviderInfosGroup,
+                    eblanApplicationInfoTags = eblanApplicationInfoTags,
+                    eblanShortcutInfosGroup = eblanShortcutInfosGroup,
+                    getEblanApplicationInfosByLabel = getEblanApplicationInfosByLabel,
+                    hasShortcutHostPermission = hasShortcutHostPermission,
+                    iconPackFilePaths = iconPackFilePaths,
+                    isPressHome = isPressHome,
+                    managedProfileResult = managedProfileResult,
+                    paddingValues = paddingValues,
+                    screenHeight = screenHeight,
+                    swipeY = swipeY,
+                    isVisibleOverlay = isVisibleOverlay,
+                    onDismiss = onDismiss,
+                    onDragEnd = onDragEnd,
+                    onDraggingGridItem = onDraggingGridItem,
+                    onEditApplicationInfo = onEditApplicationInfo,
+                    onGetEblanApplicationInfosByLabel = onGetEblanApplicationInfosByLabel,
+                    onGetEblanApplicationInfosByTagId = onGetEblanApplicationInfosByTagId,
+                    onUpdateGridItemSource = onUpdateGridItemSource,
+                    onUpdateImageBitmap = onUpdateImageBitmap,
+                    onUpdateIsDragging = onUpdateIsDragging,
+                    onUpdateOverlayBounds = onUpdateOverlayBounds,
+                    onUpdateSharedElementKey = onUpdateSharedElementKey,
+                    onVerticalDrag = onVerticalDrag,
+                    onWidgets = onWidgets,
+                    onDraggingShortcutInfoGridItem = onDraggingShortcutInfoGridItem,
+                    onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
                 )
-            },
-        )
-
-    val userHandle =
-        userManager.getUserForSerialNumber(serialNumber = eblanUserPageKey.eblanUser.serialNumber)
-
-    var isQuietModeEnabled by remember { mutableStateOf(false) }
-
-    LaunchedEffect(key1 = userHandle) {
-        if (userHandle != null) {
-            isQuietModeEnabled = userManager.isQuietModeEnabled(userHandle = userHandle)
-        }
-    }
-
-    LaunchedEffect(key1 = managedProfileResult) {
-        if (managedProfileResult != null && managedProfileResult.serialNumber == eblanUserPageKey.eblanUser.serialNumber) {
-            isQuietModeEnabled = managedProfileResult.isQuiteModeEnabled
-        }
-    }
-
-    Box(modifier = modifier.fillMaxSize()) {
-        if (isQuietModeEnabled) {
-            QuiteModeScreen(
-                packageManager = packageManager,
-                userHandle = userHandle,
-                userManager = userManager,
-                onDragEnd = onDragEnd,
-                onUpdateRequestQuietModeEnabled = { newIsQuietModeEnabled ->
-                    isQuietModeEnabled = newIsQuietModeEnabled
-                },
-                onVerticalDrag = onVerticalDrag,
-            )
-        } else if (isRearrangeEblanApplicationInfo && eblanApplicationInfoOrder == EblanApplicationInfoOrder.Index) {
-            DragAndDropEblanApplicationInfos(
-                appDrawerSettings = appDrawerSettings,
-                eblanUserPageKey = eblanUserPageKey,
-                getEblanApplicationInfosByLabel = getEblanApplicationInfosByLabel,
-                iconPackFilePaths = iconPackFilePaths,
-                paddingValues = paddingValues,
-                onDismissDragAndDrop = onDismissDragAndDrop,
-                onUpdateEblanApplicationInfos = onUpdateEblanApplicationInfos,
-            )
-        } else {
-            EblanApplicationInfos(
-                appDrawerSettings = appDrawerSettings,
-                currentPage = currentPage,
-                drag = drag,
-                eblanUserPageKey = eblanUserPageKey,
-                getEblanApplicationInfosByLabel = getEblanApplicationInfosByLabel,
-                iconPackFilePaths = iconPackFilePaths,
-                managedProfileResult = managedProfileResult,
-                paddingValues = paddingValues,
-                showPopupApplicationMenu = showPopupApplicationMenu,
-                isVisibleOverlay = isVisibleOverlay,
-                onDismiss = onDismiss,
-                onDragEnd = onDragEnd,
-                onDraggingGridItem = onDraggingGridItem,
-                onUpdateGridItemSource = onUpdateGridItemSource,
-                onUpdateImageBitmap = onUpdateImageBitmap,
-                onUpdateIsDragging = onUpdateIsDragging,
-                onUpdateOverlayBounds = onUpdateOverlayBounds,
-                onUpdatePopupMenu = onUpdatePopupMenu,
-                onUpdateSharedElementKey = onUpdateSharedElementKey,
-                onVerticalDrag = onVerticalDrag,
-                onUpdateEblanApplicationInfo = onUpdateEblanApplicationInfo,
-                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-            )
-
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P && packageManager.isDefaultLauncher() &&
-                eblanUserPageKey.eblanUser.serialNumber > 0 && userHandle != null
-            ) {
-                FloatingActionButton(
-                    modifier = Modifier
-                        .align(Alignment.BottomEnd)
-                        .padding(
-                            end = 10.dp,
-                            bottom = paddingValues.calculateBottomPadding() + 10.dp,
-                        ),
-                    onClick = {
-                        userManager.requestQuietModeEnabled(
-                            enableQuiteMode = true,
-                            userHandle = userHandle,
-                        )
-
-                        isQuietModeEnabled = userManager.isQuietModeEnabled(userHandle)
-                    },
-                ) {
-                    Icon(
-                        imageVector = EblanLauncherIcons.WorkOff,
-                        contentDescription = null,
-                    )
-                }
             }
         }
     }
 }
 
 @Composable
-private fun QuiteModeScreen(
+internal fun QuiteModeScreen(
     modifier: Modifier = Modifier,
     packageManager: AndroidPackageManagerWrapper,
     userHandle: UserHandle?,
@@ -863,182 +313,9 @@ private fun QuiteModeScreen(
     }
 }
 
-@OptIn(ExperimentalLayoutApi::class, ExperimentalSharedTransitionApi::class)
-@Composable
-private fun SharedTransitionScope.EblanApplicationInfos(
-    modifier: Modifier = Modifier,
-    appDrawerSettings: AppDrawerSettings,
-    currentPage: Int,
-    drag: Drag,
-    eblanUserPageKey: EblanUserPageKey,
-    getEblanApplicationInfosByLabel: GetEblanApplicationInfosByLabel,
-    iconPackFilePaths: Map<String, String>,
-    managedProfileResult: ManagedProfileResult?,
-    paddingValues: PaddingValues,
-    isVisibleOverlay: Boolean,
-    showPopupApplicationMenu: Boolean,
-    onDismiss: () -> Unit,
-    onDragEnd: (Float) -> Unit,
-    onDraggingGridItem: () -> Unit,
-    onUpdateGridItemSource: (GridItemSource) -> Unit,
-    onUpdateImageBitmap: (ImageBitmap) -> Unit,
-    onUpdateIsDragging: (Boolean) -> Unit,
-    onUpdateOverlayBounds: (
-        intOffset: IntOffset,
-        intSize: IntSize,
-    ) -> Unit,
-    onUpdatePopupMenu: (Boolean) -> Unit,
-    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
-    onVerticalDrag: (Float) -> Unit,
-    onUpdateEblanApplicationInfo: (EblanApplicationInfo) -> Unit,
-    onUpdateIsVisibleOverlay: (Boolean) -> Unit,
-) {
-    val scope = rememberCoroutineScope()
-
-    val overscrollEffect = remember(key1 = scope) {
-        OffsetOverscrollEffect(
-            scope = scope,
-            onVerticalDrag = onVerticalDrag,
-            onDragEnd = onDragEnd,
-        )
-    }
-
-    val lazyGridState = rememberLazyGridState()
-
-    val canScroll by remember(key1 = lazyGridState) {
-        derivedStateOf {
-            lazyGridState.canScrollForward || lazyGridState.canScrollBackward
-        }
-    }
-
-    val nestedScrollConnection = remember {
-        OffsetNestedScrollConnection(
-            onVerticalDrag = onVerticalDrag,
-            onDragEnd = onDragEnd,
-        )
-    }
-
-    var isQuietModeEnabled by remember { mutableStateOf(false) }
-
-    LaunchedEffect(key1 = lazyGridState.isScrollInProgress) {
-        if (lazyGridState.isScrollInProgress && showPopupApplicationMenu) {
-            onUpdatePopupMenu(false)
-        }
-    }
-
-    Box(
-        modifier = modifier
-            .run {
-                if (!canScroll) {
-                    nestedScroll(nestedScrollConnection)
-                } else {
-                    this
-                }
-            }
-            .fillMaxSize(),
-    ) {
-        LazyVerticalGrid(
-            columns = GridCells.Fixed(count = appDrawerSettings.appDrawerColumns),
-            state = lazyGridState,
-            modifier = Modifier.matchParentSize(),
-            contentPadding = PaddingValues(
-                bottom = paddingValues.calculateBottomPadding(),
-            ),
-            overscrollEffect = if (canScroll) {
-                overscrollEffect
-            } else {
-                rememberOverscrollEffect()
-            },
-        ) {
-            when (eblanUserPageKey.eblanUser.eblanUserType) {
-                EblanUserType.Personal -> {
-                    items(getEblanApplicationInfosByLabel.eblanApplicationInfos[eblanUserPageKey].orEmpty()) { eblanApplicationInfo ->
-                        key(eblanApplicationInfo.serialNumber, eblanApplicationInfo.componentName) {
-                            EblanApplicationInfoItem(
-                                appDrawerSettings = appDrawerSettings,
-                                currentPage = currentPage,
-                                drag = drag,
-                                eblanApplicationInfo = eblanApplicationInfo,
-                                iconPackFilePaths = iconPackFilePaths,
-                                paddingValues = paddingValues,
-                                isVisibleOverlay = isVisibleOverlay,
-                                onDismiss = onDismiss,
-                                onDraggingGridItem = onDraggingGridItem,
-                                onUpdateGridItemSource = onUpdateGridItemSource,
-                                onUpdateImageBitmap = onUpdateImageBitmap,
-                                onUpdateIsDragging = onUpdateIsDragging,
-                                onUpdateOverlayBounds = onUpdateOverlayBounds,
-                                onUpdatePopupMenu = onUpdatePopupMenu,
-                                onUpdateSharedElementKey = onUpdateSharedElementKey,
-                                onUpdateEblanApplicationInfo = onUpdateEblanApplicationInfo,
-                                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-                            )
-                        }
-                    }
-
-                    privateSpace(
-                        appDrawerSettings = appDrawerSettings,
-                        drag = drag,
-                        iconPackFilePaths = iconPackFilePaths,
-                        isQuietModeEnabled = isQuietModeEnabled,
-                        managedProfileResult = managedProfileResult,
-                        paddingValues = paddingValues,
-                        privateEblanApplicationInfos = getEblanApplicationInfosByLabel.privateEblanApplicationInfos,
-                        privateEblanUser = getEblanApplicationInfosByLabel.privateEblanUser,
-                        onUpdateIsQuietModeEnabled = { newIsQuiteModeEnabled ->
-                            isQuietModeEnabled = newIsQuiteModeEnabled
-                        },
-                        onUpdateOverlayBounds = onUpdateOverlayBounds,
-                        onUpdatePopupMenu = onUpdatePopupMenu,
-                        onUpdateEblanApplicationInfo = onUpdateEblanApplicationInfo,
-                    )
-                }
-
-                else -> {
-                    items(getEblanApplicationInfosByLabel.eblanApplicationInfos[eblanUserPageKey].orEmpty()) { eblanApplicationInfo ->
-                        key(eblanApplicationInfo.serialNumber, eblanApplicationInfo.componentName) {
-                            EblanApplicationInfoItem(
-                                appDrawerSettings = appDrawerSettings,
-                                currentPage = currentPage,
-                                drag = drag,
-                                eblanApplicationInfo = eblanApplicationInfo,
-                                iconPackFilePaths = iconPackFilePaths,
-                                paddingValues = paddingValues,
-                                isVisibleOverlay = isVisibleOverlay,
-                                onDismiss = onDismiss,
-                                onDraggingGridItem = onDraggingGridItem,
-                                onUpdateGridItemSource = onUpdateGridItemSource,
-                                onUpdateImageBitmap = onUpdateImageBitmap,
-                                onUpdateIsDragging = onUpdateIsDragging,
-                                onUpdateOverlayBounds = onUpdateOverlayBounds,
-                                onUpdatePopupMenu = onUpdatePopupMenu,
-                                onUpdateSharedElementKey = onUpdateSharedElementKey,
-                                onUpdateEblanApplicationInfo = onUpdateEblanApplicationInfo,
-                                onUpdateIsVisibleOverlay = onUpdateIsVisibleOverlay,
-                            )
-                        }
-                    }
-                }
-            }
-        }
-
-        if (!WindowInsets.isImeVisible) {
-            ScrollBarThumb(
-                modifier = Modifier
-                    .align(Alignment.TopEnd)
-                    .fillMaxHeight(),
-                appDrawerSettings = appDrawerSettings,
-                lazyGridState = lazyGridState,
-                paddingValues = paddingValues,
-                onScrollToItem = lazyGridState::scrollToItem,
-            )
-        }
-    }
-}
-
 @OptIn(ExperimentalUuidApi::class, ExperimentalSharedTransitionApi::class)
 @Composable
-private fun SharedTransitionScope.EblanApplicationInfoItem(
+internal fun SharedTransitionScope.EblanApplicationInfoItem(
     modifier: Modifier = Modifier,
     appDrawerSettings: AppDrawerSettings,
     currentPage: Int,
@@ -1284,7 +561,7 @@ private fun SharedTransitionScope.EblanApplicationInfoItem(
 
 @Composable
 @OptIn(ExperimentalMaterial3Api::class)
-private fun EblanApplicationInfoTabRow(
+internal fun EblanApplicationInfoTabRow(
     modifier: Modifier = Modifier,
     currentPage: Int,
     eblanApplicationInfos: Map<EblanUserPageKey, List<EblanApplicationInfo>>,
@@ -1316,7 +593,7 @@ private fun EblanApplicationInfoTabRow(
 }
 
 @Composable
-private fun TagElevatedFilterChip(
+internal fun TagElevatedFilterChip(
     modifier: Modifier = Modifier,
     eblanApplicationInfoTag: EblanApplicationInfoTag,
     selectedEblanApplicationInfoTag: Long?,
