+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/application/vertical/LazyGridDragAndDrop.kt
@@ -15,7 +15,7 @@
  *   limitations under the License.
  *
  */
-package com.eblan.launcher.feature.home.screen.application.draganddrop
+package com.eblan.launcher.feature.home.screen.application.vertical
 
 import androidx.compose.animation.core.Animatable
 import androidx.compose.animation.core.Spring
